Add bc compatibility to RestController by adding new AbstractRestController
