+++ src/Sulu/Bundle/ContactBundle/Controller/ContactController.php
@@ -31,7 +31,7 @@ use Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor\DoctrineJoinDescrip
 use Sulu\Component\Rest\ListBuilder\FieldDescriptorInterface;
 use Sulu\Component\Rest\ListBuilder\ListRepresentation;
 use Sulu\Component\Rest\ListBuilder\Metadata\FieldDescriptorFactoryInterface;
-use Sulu\Component\Rest\RestController;
+use Sulu\Component\Rest\AbstractRestController;
 use Sulu\Component\Rest\RestHelperInterface;
 use Sulu\Component\Security\Authentication\UserRepositoryInterface;
 use Sulu\Component\Security\SecuredControllerInterface;
@@ -41,7 +41,7 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInt
 /**
  * Makes contacts available through a REST API.
  */
-class ContactController extends RestController implements ClassResourceInterface, SecuredControllerInterface
+class ContactController extends AbstractRestController implements ClassResourceInterface, SecuredControllerInterface
 {
     /**
      * {@inheritdoc}
