+++ src/Sulu/Bundle/CoreBundle/Controller/LocalizationController.php
@@ -15,13 +15,13 @@ use FOS\RestBundle\Routing\ClassResourceInterface;
 use FOS\RestBundle\View\ViewHandlerInterface;
 use Sulu\Component\Localization\Manager\LocalizationManagerInterface;
 use Sulu\Component\Rest\ListBuilder\CollectionRepresentation;
-use Sulu\Component\Rest\RestController;
+use Sulu\Component\Rest\AbstractRestController;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
  * Controller which returns the localizations for the entire system.
  */
-class LocalizationController extends RestController implements ClassResourceInterface
+class LocalizationController extends AbstractRestController implements ClassResourceInterface
 {
     /**
      * @var LocalizationManagerInterface
