+++ src/Sulu/Bundle/MediaBundle/Controller/FormatController.php
@@ -16,10 +16,10 @@ use FOS\RestBundle\View\ViewHandlerInterface;
 use Sulu\Bundle\MediaBundle\Media\FormatManager\FormatManagerInterface;
 use Sulu\Component\Rest\ListBuilder\CollectionRepresentation;
 use Sulu\Component\Rest\RequestParametersTrait;
-use Sulu\Component\Rest\RestController;
+use Sulu\Component\Rest\AbstractRestController;
 use Symfony\Component\HttpFoundation\Request;
 
-class FormatController extends RestController implements ClassResourceInterface
+class FormatController extends AbstractRestController implements ClassResourceInterface
 {
     use RequestParametersTrait;
