+++ src/Sulu/Bundle/PageBundle/Controller/VersionController.php
@@ -24,7 +24,7 @@ use Sulu\Component\Rest\ListBuilder\ListRestHelper;
 use Sulu\Component\Rest\ListBuilder\ListRestHelperInterface;
 use Sulu\Component\Rest\ListBuilder\Metadata\FieldDescriptorFactoryInterface;
 use Sulu\Component\Rest\RequestParametersTrait;
-use Sulu\Component\Rest\RestController;
+use Sulu\Component\Rest\AbstractRestController;
 use Sulu\Component\Rest\RestHelperInterface;
 use Sulu\Component\Security\Authentication\UserRepositoryInterface;
 use Sulu\Component\Security\Authorization\AccessControl\SecuredObjectControllerInterface;
@@ -38,7 +38,7 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInt
 /**
  * Handles the versions of pages.
  */
-class VersionController extends RestController implements
+class VersionController extends AbstractRestController implements
     ClassResourceInterface,
     SecuredControllerInterface,
     SecuredObjectControllerInterface
