+++ src/Sulu/Bundle/SecurityBundle/Controller/ContextsController.php
@@ -14,10 +14,10 @@ namespace Sulu\Bundle\SecurityBundle\Controller;
 use FOS\RestBundle\Routing\ClassResourceInterface;
 use FOS\RestBundle\View\ViewHandlerInterface;
 use Sulu\Bundle\AdminBundle\Admin\AdminPool;
-use Sulu\Component\Rest\RestController;
+use Sulu\Component\Rest\AbstractRestController;
 use Symfony\Component\HttpFoundation\Request;
 
-class ContextsController extends RestController implements ClassResourceInterface
+class ContextsController extends AbstractRestController implements ClassResourceInterface
 {
     /**
      * @var AdminPool
