+++ src/Sulu/Component/Rest/Tests/Unit/AbstractRestControllerTest.php
@@ -16,9 +16,9 @@ use PHPUnit\Framework\TestCase;
 use Sulu\Bundle\CoreBundle\Entity\ApiEntity;
 use Sulu\Component\Rest\Exception\EntityNotFoundException;
 use Sulu\Component\Rest\Exception\RestException;
-use Sulu\Component\Rest\RestController;
+use Sulu\Component\Rest\AbstractRestController;
 
-class RestControllerTest extends TestCase
+class AbstractRestControllerTest extends TestCase
 {
     /**
      * @var \Sulu\Component\Rest\RestController
@@ -27,12 +27,12 @@ class RestControllerTest extends TestCase
 
     public function setUp(): void
     {
-        $this->controller = $this->getMockForAbstractClass(RestController::class);
+        $this->controller = $this->getMockForAbstractClass(AbstractRestController::class);
     }
 
     public function testResponseGetById()
     {
-        $method = new \ReflectionMethod(RestController::class, 'responseGetById');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'responseGetById');
         $method->setAccessible(true);
 
         $id = 1;
@@ -49,7 +49,7 @@ class RestControllerTest extends TestCase
 
     public function testResponseGetByNotExistingId()
     {
-        $method = new \ReflectionMethod(RestController::class, 'responseGetById');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'responseGetById');
         $method->setAccessible(true);
 
         $id = 1;
@@ -77,7 +77,7 @@ class RestControllerTest extends TestCase
             $this->fail('add should not be called');
         };
 
-        $method = new \ReflectionMethod(RestController::class, 'processPut');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'processPut');
         $method->setAccessible(true);
 
         $result = $method->invoke($this->controller, [], [], $delete, $update, $add);
@@ -99,7 +99,7 @@ class RestControllerTest extends TestCase
             $this->fail('add should not be called');
         };
 
-        $method = new \ReflectionMethod(RestController::class, 'processPut');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'processPut');
         $method->setAccessible(true);
 
         $object = $this->prophesize(ApiEntity::class);
@@ -134,7 +134,7 @@ class RestControllerTest extends TestCase
             $this->fail('add should not be called');
         };
 
-        $method = new \ReflectionMethod(RestController::class, 'processPut');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'processPut');
         $method->setAccessible(true);
 
         $object = $this->prophesize(ApiEntity::class);
@@ -173,7 +173,7 @@ class RestControllerTest extends TestCase
             $addCalled = true;
         };
 
-        $method = new \ReflectionMethod(RestController::class, 'processPut');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'processPut');
         $method->setAccessible(true);
 
         $method->invoke(
@@ -194,7 +194,7 @@ class RestControllerTest extends TestCase
 
     public function testDelete()
     {
-        $method = new \ReflectionMethod(RestController::class, 'responseDelete');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'responseDelete');
         $method->setAccessible(true);
 
         $id = 1;
@@ -211,7 +211,7 @@ class RestControllerTest extends TestCase
 
     public function testDeleteWithNotExistingEntity()
     {
-        $method = new \ReflectionMethod(RestController::class, 'responseDelete');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'responseDelete');
         $method->setAccessible(true);
 
         $id = 1;
@@ -227,7 +227,7 @@ class RestControllerTest extends TestCase
 
     public function testDeleteWithError()
     {
-        $method = new \ReflectionMethod(RestController::class, 'responseDelete');
+        $method = new \ReflectionMethod(AbstractRestController::class, 'responseDelete');
         $method->setAccessible(true);
 
         $id = 1;
